requirements.txt
Package            Version
------------------ ---------
aiogram            2.25.1
aiohttp            3.8.4
aiosignal          1.3.1
async-timeout      4.0.2
attrs              23.1.0
Babel              2.9.1
certifi            2022.12.7
charset-normalizer 3.1.0
frozenlist         1.3.3
geojson            2.5.0
idna               3.4
magic-filter       1.0.9
multidict          6.0.4
pip                22.3.1
pyowm              3.3.0
PySocks            1.7.1
python-dotenv      1.0.0
pytz               2023.3
requests           2.28.2
setuptools         65.5.1
urllib3            1.26.15
wheel              0.38.4
yarl               1.8.2
